Add os-conf-acceptance-tests suite
This change adds a basic test suite to test the functionality of every
job in this release.

The test suite operates on an existing bosh deployment. It currently
deploys an example manifest with every os-conf-release job colocated on
a single VM. It will then use the bosh cli to validate that the jobs are
operating correctly.

Currently we do not test the `disable_agent` job as the resulting
unresponsive agent will make the acceptance tests slow.

Also bosh-lite is unsupported as many of the os-conf-release jobs are no
applicable.

[#158400410](https://www.pivotaltracker.com/story/show/158400410)

Co-authored-by: Michael Xu <mxu@pivotal.io>