🐞 Assume `systemd` is running; don't check
As of Bionic & later, `systemd` is always running; we don't need to check for
it in the pre-start script; it invites spurious errors.

`systemd` replaces `init` (PID 1, the grandparent of all processes). It's the
first process run, and it's always running, and you can't kill it, `-9`
notwithstanding.

Newer Bionic stemcells masquerade `systemd` as `/sbin/init` for
backwards-compatibility. It also spawns a separate `systemd` later in the boot
process.  This introduces a race condition in our pre-start script, which
checks if `systemd` is running: it doesn't always find a process named `systemd`

This commit removes the check for `systemd`.

Brian Cunnie assures, "`systemd` will _always_ be running."

fixes:
```
Limits
/tmp/build/bd5f12f4/os-conf-release/src/os-conf-acceptance-tests/limits_test.go:12
  when limits are configured
  /tmp/build/bd5f12f4/os-conf-release/src/os-conf-acceptance-tests/limits_test.go:19
    sets the limits for the systemd process [It]
    /tmp/build/bd5f12f4/os-conf-release/src/os-conf-acceptance-tests/limits_test.go:40

    No future change is possible.  Bailing out early after 1.918s.
    Got stuck at:
        os-conf/6ae65a92-263c-4f36-a70e-5dbeefc8963d: stderr | Jim was here.
        os-conf/6ae65a92-263c-4f36-a70e-5dbeefc8963d: stdout | 1024
        os-conf/6ae65a92-263c-4f36-a70e-5dbeefc8963d: stderr | Connection to 10.0.0.4 closed.

    Waiting for:
        60000
```

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>