add jammy compability
the harden sshd job wouldn't work on jammy because it uses a newer
openssh version than 7. the initial commit that updated the behaviour to
add trusty support

see: 3d44807717847bb9ddca4de9f7d9348ecee9064e

on jammy that check (which I assume is checking that the minimum version
is OpenSSH_7) fails and the property is not added. Update the check to
parse out the major version and check if is greater than 6 which should
keep trusty compability and work for future two digit major versions of
OpenSSH.

additionally the test deployments tried loading aufs which is shipped
with jammy anymore. so loading it won't work either. switch to `lp` kernel
module because lp is also used in the stemcell builder tests and is not
only available in bionic and jammy but also in fips mode (which for this
usecase doesn't matter).

compare:

https://github.com/cloudfoundry/bosh-linux-stemcell-builder/commit/19b512a91aa65c619e204deb4eb8b2d53b6f118a